XenServer installer script
Related to XenServer/GateIntegration:
    https://wiki.openstack.org/wiki/XenServer/GateIntegration

convert_node_to_xenserver.sh will convert a Rackspace image to a
XenServer 6.2 with an appliance on top of it. install_xenserver.sh is
calling the script with the appropriate parameters.

The incubator project for `convert_node_to_xenserver.sh`:
    https://github.com/citrix-openstack/xenapi-in-the-cloud

The project page for generating the appliance:
    https://github.com/citrix-openstack/openstack-xenapi-testing-xva

As the script is finished, the appliance will be accessible on the
instance's public IP. Dom0 could be reached from domU on the IP:
192.168.32.2

Flavor:
    performance1-8
Image:
    62df001e-87ee-407c-b042-6f4e13f5d7e1

Change-Id: Ibe94e56f6f3ac890e7f269042bc1db525b413efe